refactor: move commandData files into one file called AuxdibotCommandData
