+++ lib/src/bot/commands/embeds/embed.ts
@@ -15,7 +15,7 @@ import { toAPIEmbed } from '@util/types/EmbedParameters';
 import { getMessage } from '@util/functions/getMessage';
 import parsePlaceholders from '@util/functions/parsePlaceholder';
 import AuxdibotCommandInteraction from '@util/templates/AuxdibotCommandInteraction';
-import GuildAuxdibotCommandData from '@util/types/commandData/GuildAuxdibotCommandData';
+import { GuildAuxdibotCommandData } from '@util/types/AuxdibotCommandData';
 import createEmbedParameters from '@util/functions/createEmbedParameters';
 import argumentsToEmbedParameters from '@util/functions/argumentsToEmbedParameters';
