+++ lib/src/bot/commands/levels/levels.ts
@@ -1,7 +1,7 @@
 import { PermissionsBitField, SlashCommandBuilder } from 'discord.js';
 import AuxdibotCommand from '@util/templates/AuxdibotCommand';
 import AuxdibotCommandInteraction from '@util/templates/AuxdibotCommandInteraction';
-import GuildAuxdibotCommandData from '@util/types/commandData/GuildAuxdibotCommandData';
+import { GuildAuxdibotCommandData } from '@util/types/AuxdibotCommandData';
 import Embeds from '@util/constants/Embeds';
 import { ILevelReward } from '@schemas/LevelRewardSchema';
 const levelCommand = <AuxdibotCommand>{
