+++ lib/src/bot/commands/levels/mylevel.ts
@@ -3,7 +3,7 @@ import AuxdibotCommand from '@util/templates/AuxdibotCommand';
 import Embeds from '@util/constants/Embeds';
 import dotenv from 'dotenv';
 import AuxdibotCommandInteraction from '@util/templates/AuxdibotCommandInteraction';
-import GuildAuxdibotCommandData from '@util/types/commandData/GuildAuxdibotCommandData';
+import { GuildAuxdibotCommandData } from '@util/types/AuxdibotCommandData';
 import calcXP from '@util/functions/calcXP';
 dotenv.config();
 const myLevelCommand = <AuxdibotCommand>{
