+++ lib/src/bot/commands/permissions/permissions.ts
@@ -3,7 +3,7 @@ import AuxdibotCommand from '@util/templates/AuxdibotCommand';
 import Embeds from '@util/constants/Embeds';
 import { IPermissionOverride } from '@schemas/PermissionOverrideSchema';
 import AuxdibotCommandInteraction from '@util/templates/AuxdibotCommandInteraction';
-import GuildAuxdibotCommandData from '@util/types/commandData/GuildAuxdibotCommandData';
+import { GuildAuxdibotCommandData } from '@util/types/AuxdibotCommandData';
 import { LogType } from '@util/types/Log';
 
 const permissionsCommand = <AuxdibotCommand>{
