+++ lib/src/bot/commands/suggestions/suggestions.ts
@@ -1,7 +1,7 @@
 import { APIEmbed, ChannelType, GuildBasedChannel, SlashCommandBuilder } from 'discord.js';
 import AuxdibotCommand from '@util/templates/AuxdibotCommand';
 import AuxdibotCommandInteraction from '@util/templates/AuxdibotCommandInteraction';
-import GuildAuxdibotCommandData from '@util/types/commandData/GuildAuxdibotCommandData';
+import { GuildAuxdibotCommandData } from '@util/types/AuxdibotCommandData';
 import SuggestionState, { SuggestionStateName } from '@util/types/SuggestionState';
 import Embeds from '@util/constants/Embeds';
 import { ISuggestion } from '@schemas/SuggestionSchema';
