+++ lib/src/util/templates/AuxdibotCommandInteraction.ts
@@ -1,5 +1,5 @@
 import { ChatInputCommandInteraction } from 'discord.js';
-import BaseAuxdibotCommandData from '../types/commandData/BaseAuxdibotCommandData';
+import { BaseAuxdibotCommandData } from '../types/AuxdibotCommandData';
 
 export default interface AuxdibotCommandInteraction<Data extends BaseAuxdibotCommandData>
    extends ChatInputCommandInteraction {
