--- lib/src/util/types/commandData/BaseAuxdibotCommandData.ts
@@ -1,4 +1,0 @@
-export default interface BaseAuxdibotCommandData {
-   dmCommand?: boolean;
-   date: Date;
-}
