--- lib/src/util/types/commandData/DMAuxdibotCommandData.ts
@@ -1,8 +1,0 @@
-import { User } from 'discord.js';
-import BaseAuxdibotCommandData from './BaseAuxdibotCommandData';
-
-export default interface DMAuxdibotCommandData extends BaseAuxdibotCommandData {
-   dmCommand?: true;
-   date: Date;
-   user: User;
-}
